refactor: remove duplicate and collapsable imports
