+++ plugin-hrm-form/src/components/case/ViewCaseItem.tsx
@@ -21,8 +21,7 @@ import { Template } from '@twilio/flex-ui';
 import Edit from '@material-ui/icons/Edit';
 import { DefinitionVersion, isNonSaveable } from 'hrm-form-definitions';
 
-import { BottomButtonBar, Box } from '../../styles';
-import { StyledNextStepButton } from '../../styles/buttons';
+import { BottomButtonBar, Box, StyledNextStepButton } from '../../styles';
 import { CaseLayout, FullWidthFormTextContainer } from './styles';
 import { RootState } from '../../states';
 import { SectionEntry, SectionEntryValue } from '../common/forms/SectionEntry';
