+++ plugin-hrm-form/src/components/caseList/CaseList.tsx
@@ -22,8 +22,7 @@ import { DefinitionVersion } from 'hrm-form-definitions';
 import Case from '../case';
 import { StandaloneITask, Case as CaseType } from '../../types/types';
 import CaseListTable from './CaseListTable';
-import { ListContainer, CenteredContainer, SomethingWentWrongText } from '../../styles/table';
-import { listCases } from '../../services/CaseService';
+import { ListContainer, CenteredContainer, SomethingWentWrongText } from '../../styles';
 import { CaseLayout } from '../case/styles';
 import * as ConfigActions from '../../states/configuration/actions';
 import { StandaloneSearchContainer } from '../search/styles';
