+++ plugin-hrm-form/src/components/caseList/CaseListTableRow.tsx
@@ -35,8 +35,9 @@ import {
   IdentifierContainer,
   IdentifierActionButton,
   TableSummaryFont,
-} from '../../styles/table';
-import { Box, HiddenText } from '../../styles';
+  Box,
+  HiddenText,
+} from '../../styles';
 import { formatName, getShortSummary } from '../../utils';
 import { getContactTags } from '../../utils/categories';
 import CategoryWithTooltip from '../common/CategoryWithTooltip';
