+++ plugin-hrm-form/src/components/caseList/filters/SearchInput.tsx
@@ -20,7 +20,7 @@ import React, { Ref } from 'react';
 import SearchIcon from '@material-ui/icons/Search';
 import ClearIcon from '@material-ui/icons/Clear';
 
-import { MultiSelectSearchInput } from '../../../styles/filters';
+import { MultiSelectSearchInput } from '../../../styles';
 
 type OwnProps = {
   label: string;
